python/dist-langchain/examples/multi_agent_return_eligibility/requirements.txt
alation-ai-agent-sdk~=0.1.2
alation-ai-agent-langchain~=0.1.1
langchain>=0.1.0
langchain-openai>=0.0.2
langgraph>=0.0.20
python-dotenv>=1.0.0
pydantic>=2.0.0
